Update schema and delete old migrations
