--- drizzle/0000_outstanding_nekra.sql
@@ -1,5 +1,0 @@
-CREATE TABLE `todos` (
-	`id` integer PRIMARY KEY NOT NULL,
-	`content` text,
-	`done` integer
-);
