--- drizzle/meta/_journal.json
@@ -1,13 +1,0 @@
-{
-  "version": "5",
-  "dialect": "sqlite",
-  "entries": [
-    {
-      "idx": 0,
-      "version": "5",
-      "when": 1696905863833,
-      "tag": "0000_outstanding_nekra",
-      "breakpoints": true
-    }
-  ]
-}
