--- src/server/trpc.ts
@@ -1,6 +1,0 @@
-import { initTRPC } from '@trpc/server';
-
-const t = initTRPC.create();
-
-export const router = t.router;
-export const publicProcedure = t.procedure;
